fix warnings

+++ stepmania/src/CourseUtil.cpp
@@ -150,7 +150,7 @@ void CourseUtil::SortCoursePointerArrayByAvgDifficulty( vector<Course*> &apCours
 	for(unsigned i = 0; i < apCourses.size(); ++i)
 	{
 		Trail* pTrail = apCourses[i]->GetTrail( GAMESTATE->GetCurrentStyleDef()->m_StepsType, COURSE_DIFFICULTY_REGULAR );
-		course_sort_val[apCourses[i]] = pTrail->GetMeter();
+		course_sort_val[apCourses[i]] = (float) pTrail->GetMeter();
 	}
 	sort( apCourses.begin(), apCourses.end(), CompareCoursePointersByTitle );
 	stable_sort( apCourses.begin(), apCourses.end(), CompareCoursePointersBySortValueAscending );
